% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latinr_article.R
\name{latinr_article}
\alias{latinr_article}
\title{R Markdown output format for submissions to LatinR}
\usage{
latinr_article(
  ...,
  keep_tex = FALSE,
  highlight = "default",
  citation_package = "default",
  latex_engine = "xelatex",
  anonymise = TRUE,
  author_bottom = FALSE
)
}
\arguments{
\item{...}{Arguments passed to \code{pdf_document()}.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", "xelatex" and "tectonic".}

\item{anonymise}{whether to anonymise the PDF}

\item{author_bottom}{whether to add author information at the bottom of the
page.}
}
\description{
R Markdown output format for submissions to LatinR
}
