% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latinr_article.R
\name{latinr_checks}
\alias{latinr_checks}
\title{Checks that metadata is ok}
\usage{
latinr_checks(metadata, check_is_error = TRUE)
}
\arguments{
\item{metadata}{yaml metadata.}

\item{check_is_error}{whether to treat fails as errors or only warnings
useful during development).}
}
\description{
Does basics checks on metadata so that it adheres to submission guidelines.
}
