% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{latinr_default_user}
\alias{latinr_default_user}
\alias{latinr_default_user_get}
\alias{latinr_default_user_set}
\title{Default users}
\usage{
latinr_default_user_set(user = NULL)

latinr_default_user_get()
}
\arguments{
\item{user}{user to use as the default (`NULL` means no default user).}
}
\value{
`latinr_default_user_get()` returns the default user set with `latinr_default_user_get()`
or the user if there's only one user saved with `latinr_password_set()`.

`latinr_default_user_get()` sets the default user (for the current session) 
and returns the supplied user invisibly.
}
\description{
Default users
}
\section{Functions}{
\itemize{
\item \code{latinr_default_user_set}: Sets the default user.

\item \code{latinr_default_user_get}: Gets the default user.
}}

