% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{latinr_submit}
\alias{latinr_submit}
\title{Submit an article to LatinR}
\usage{
latinr_submit(rmd = list.files(getwd(), pattern = ".Rmd"),
  user = latinr_default_user_get(), user_check = TRUE)
}
\description{
Submit an article to LatinR
}
