% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latinr-package.R
\docType{package}
\name{latinr-package}
\alias{latinr}
\alias{latinr-package}
\title{latinr: Submission to LatinR}
\description{
Facilitates creating submissions to the LatinR confference.
}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@fcen.cima.fcen.uba.ar}

}
\keyword{internal}
