% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latinr_article.R
\name{latinr_article}
\alias{latinr_article}
\title{R Markdown output format for submissions to LatinR}
\usage{
latinr_article(..., keep_tex = TRUE, citation_package = "natbib",
  highlight = NULL, md_extensions = c("-autolink_bare_uris",
  "-auto_identifiers"))
}
\description{
R Markdown output format for submissions to LatinR
}
