% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{latinr_lists}
\alias{latinr_lists}
\alias{latinr_topics}
\alias{latinr_countries}
\title{Latinr lists}
\usage{
latinr_topics()

latinr_countries()
}
\description{
Lists of important information for filling in metadata.
}
\section{Functions}{
\itemize{
\item \code{latinr_topics}: List of topics

\item \code{latinr_countries}: List of countries with country codes.
}}

